% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backFillNA.R
\name{backFillNA}
\alias{backFillNA}
\title{A function to back fill NAs in a vector with the last non-NA value}
\usage{
backFillNA(x)
}
\arguments{
\item{x}{A vector with some NAs}
}
\value{
A vector with NAs back filled
}
\description{
The function will throw errors when the input is not an atomic vector.
}
\details{
see https://gist.github.com/garyfeng/27e7f8e406192a8cb33a
}

