% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cenretss.r
\docType{data}
\name{cenretss}
\alias{cenretss}
\title{Census retirement state summaries.}
\format{Data frame with 1 row per state per administration level per variable per year.}
\source{
U.S. Bureau of the Census
}
\usage{
cenretss
}
\description{
State-level summaries of Census Bureau Retirement Survey, selected years 1957-1977, annual 1982-2013.
}
\examples{
library(dplyr)
glimpse(cenretss)
}
\keyword{datasets}

