% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkDfCols.R
\name{checkDfCols}
\alias{checkDfCols}
\title{Function that checks whether a set of columns given as a vector of strings is in a data frame and, if not, gives feedback on
  which are missing before stopping execution of the program.}
\usage{
checkDfCols(requiredCols, dfName,
  specifiedIn = "which was specified in preferences")
}
\arguments{
\item{requiredCols}{A character vector containg the names of columns the presence of which in a data frame is to be verfied}

\item{dfName}{Name of the data frame in which the columns in requiredCols are to be found}

\item{specifiedIn}{An optional argument (default) that allows the output to detail where the varaible names can be changed by the user}
}
\value{
TRUE, unless it stopped
}
\description{
Function that checks whether a set of columns given as a vector of strings is in a data frame and, if not, gives feedback on
  which are missing before stopping execution of the program.
}

