% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkPrefs.R
\name{checkPrefs}
\alias{checkPrefs}
\title{Function that checks user preferences for heatmapanalysis.R, and, where necessary, overwrites with default preferences and displays
a message to the user detailing the problematic preference specification and the default value used.
Problems with incorrectly specified variable names lead to the program being terminated rather than the use of defaults.}
\usage{
checkPrefs(prefs, dprefs)
}
\arguments{
\item{prefs}{A list of preferences for use in heatmapanalysis.R}

\item{dprefs}{A list of defaults for the preferences for use in heatmapanalysis.R}
}
\value{
prefs A list of preferences for use in heatmapanalysis.R in which the entries have been check and, where neede, replaced
          with default values
}
\description{
Function that checks user preferences for heatmapanalysis.R, and, where necessary, overwrites with default preferences and displays
a message to the user detailing the problematic preference specification and the default value used.
Problems with incorrectly specified variable names lead to the program being terminated rather than the use of defaults.
}

