% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_plot.R
\name{create_fip_frame}
\alias{create_fip_frame}
\title{Create FIP Frame}
\usage{
create_fip_frame(fipData)
}
\arguments{
\item{fipData}{A data frame with two columns that has
the FIP values and corresponding frequency counts.}
}
\description{
A function that transforms a data frame
 into one that is used by choroplethr.
}
