% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create Map}
\usage{
create_map(myZipcodes)
}
\arguments{
\item{myZipCodes}{A vector of integer values that _should_
represent US state zip codes.}
}
\description{
A function that takes a vector of zip codes
 and generates a hashmap of zip codes to US states. This
 function also creates a map of the US with frequency counts
 of the zipCodes in each state.
}
