% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gazeHeatmap.R
\name{gazeHeatmap}
\alias{gazeHeatmap}
\title{This function reads in a picture file specified by the user, then overlays it with a heat map computed according to user
  specifications.}
\usage{
gazeHeatmap(dframe, pict, path, outputtitle = "User Output",
  colormin = "white", colormax = "red4", screendim = c(1680, 1050),
  transparancyLevel = 0.4, ngridpoints = 100, showFixations = F)
}
\arguments{
\item{dframe}{Data frame containing the data to be used, with at least $x, $y, and $dur}

\item{pict}{String giving the base name of the PNG file to be used, e.g., "background.png"}

\item{path}{String giving the path of the PNG file to be used, e.g., path=getwd()}

\item{outputtitle}{Optional argument, allowing the user to specify the output title}

\item{colormin}{Optional argument, allowing the user to specify the color corresponding to the lowest value of the heat map}

\item{colormax}{Optional argument, allowing the user to specify the color corresponding to the highest value of the heat map}

\item{screendim}{Optional argument, allowing the user to specify the dimensions of the screen that was viewed}

\item{transparancyLevel}{Optional argument, allowing the user to set the transparancy level of the heat map}

\item{ngridpoints}{Optional argument, allowing the user to set the resolution of the heat map}

\item{showFixations}{Optional argument, to show individual fixations if TRUE; default to FALSE for heat map}
}
\value{
ggplot object: heat map or fixation map of gaze time overlayed onto a picture read from file
}
\description{
This function reads in a picture file specified by the user, then overlays it with a heat map computed according to user
  specifications.
}
\examples{
gazeHeatmap(tmp, paste(storyName, " Summary", sep=""), path=getwd(),
    screendim = screenSize, transparancyLevel = alphaLevel, ngridpoints = 200,
    showFixations = F)
}

