% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gazeTimeHist.R
\name{gazeTimeHist}
\alias{gazeTimeHist}
\title{plot distribution of gaze time, in seconds (assuming input in ms), by subject and by task
   user can specify input variables subject, time variable, task variable as well as title, subject variable lable and task variable label}
\usage{
gazeTimeHist(df1, SubVar, TimeVar, TaskVar, MainTitle = "User Output",
  SubVarLabel = "Subject ID", TaskLabel = "Task")
}
\arguments{
\item{df1}{Input data frame that contains the variables specified by the arguments below, so at the very least three variables: a
variable with a subject ID, a variable with the gaze time to be plotted, and variable with a Task ID}

\item{SubVar}{The name of the variable in df1 that contains the subject ID}

\item{TimeVar}{The name of the variable in df1 that contains the gaze time}

\item{TaskVar}{The name of the varaible in df1 that containts the task variable}

\item{MainTitle}{Optional argument, allows the user to set the title printed above the histogram}

\item{SubVarLabel}{Optional argument, allows the user to set the label of the subject variable (visible in plot legend)}

\item{TaskLabel}{Optinoal argument, allows the user to set the label of the Task variable (X-axis label)}
}
\value{
ggplot object: histogram of gaze time, by subject and task
}
\description{
plot distribution of gaze time, in seconds (assuming input in ms), by subject and by task
   user can specify input variables subject, time variable, task variable as well as title, subject variable lable and task variable label
}
\examples{
gazeTimeHist (df1, SubVar, TimeVar, TaskVar,
    MainTitle="User Output", SubVarLabel="Subject ID", TaskLabel="Task")
}

