% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uni_states.R
\name{get_uni_states}
\alias{get_uni_states}
\title{Get Unique States}
\usage{
get_uni_states(allStates)
}
\arguments{
\item{allStates}{A character vector of (possible) US
states.}
}
\description{
A function that takes a vector of multiple
 states and calculates the unique possibilities, while
 filtering out states that don't comply.
}
