% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDistr.R
\name{getDistr}
\alias{getDistr}
\title{To compute the empirical density function and hazard rate of response time}
\usage{
getDistr(df, timeVar = "t", binwidth = 50)
}
\arguments{
\item{df}{The observed response time, either as a vector or a data frame}

\item{timeVar}{(Optional) The name of the RT variable if df is a data frame
Useful when using the pipeline.}

\item{binwidth}{(optional) The width of the bin used to computer statistics; in msec.}
}
\value{
a data frame of distributional statistics such as the pdf, hazard rate, counts, etc.
}
\description{
Human response time, often measured in milliseconds, follow lawful distributions.
Given a sample of response time (as a vector or a variable in a data frame),
calculate the pdf and hazard rate along with other statistics using the
library "muhaz" and other methods.
}

