% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getExtInfo.R
\name{getExtInfo}
\alias{getExtInfo}
\title{A function to map the extendedInfo into a list of lists}
\usage{
getExtInfo(nodes)
}
\arguments{
\item{nodes}{A set of XML nodes from getNodeSet() that point to '//observableData/observableDatum'}
}
\value{
a list of lists, with names of the list elements set to the "key" and value to "value"
}
\description{
A function to map the extendedInfo into a list of lists
}
\details{
This assumes a certain structure of the XML, i.e.,

  \code{//observableData/observableDatum/content/pair/key}

  \code{//observableData/observableDatum/content/pair/value}

  where \code{key} and \code{value} are automic class, i.e., numeric or string. The code then
  makes a list with named elements, i.e., 'key' as the key, and 'value' as the value in the list.
}

