% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gfDiagrammR.R
\name{gf_makeGraphFromSubgraphs}
\alias{gf_makeGraphFromSubgraphs}
\title{Great a DOT graph from subgraphs}
\usage{
gf_makeGraphFromSubgraphs(...)
}
\arguments{
\item{...}{objects that are subgraphs created using gf_createSubgraph(). This can be individually listed
subgraphs, or a list with subgraphs}
}
\value{
DOT of the graph with all the subgraphs
}
\description{
Great a DOT graph from subgraphs
}

