% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmapanalysis.R
\name{heatmapanalysis}
\alias{heatmapanalysis}
\title{Main function of GFEye package. Gets user preferences, check user input, calls other programs to create output.}
\usage{
heatmapanalysis(datafilename, picturelocation, outfilename, outputTitle,
  preferences, coverPage = "On")
}
\arguments{
\item{datafilename}{String with valid path and location for input data. Input data must conform to this scheme:
1 header line, at least 12 columns of data:
 subjectID, vtime, gazetime, x, y, page name, AOI_ID, AOI_content, x1, y1, x2, y2}

\item{picturelocation}{Valid path of folder containing picture files over which the heat maps are to be placed. These
must be .png files, and the file names must correspond to the page name in datafilenames}

\item{outfilename}{Optional if no preferences, name of output file. If this does not contain a path, the file with
be saved in the default directory.}

\item{preferences}{Optional, user preferences for heat map. Must be either an R list or a file (with header line)
containing
1) string indicating a valid R color for the lowest value of the heat map
2) string indicating a valid R c color of the highest value of the heat map
3) vector of screen size to be used (in file: two entries, not a vector)
4) numeric value between 0 and 1 transparancy value for heat map
5) numberic value indicating the desired resolution for heat map
6) string with name of the variable containing the subject ID
7) string with name of the variable containing gaze time
8) string with name of the variable containing AOI ID
9) string with name of the variable containing x Coordinates
10) string with name of the variable containing y Coordinates
11) string with label for the Subject ID variable
12) string with label for the AOI ID variable
Any argument that is not followed by other arguments is optional.
If read from a file, the file must be comma-sepearated with no headers.}

\item{coverpage}{Optional, an argument that allows the user to turn off the cover page of the report. Default="On".
                Any user-supplied value other than the exact string "On" will turn cover page genration off.}
}
\value{
PDF file named according to outputfilename containg a summary of gaze time by AOI and subject as well as heat maps
}
\description{
The first 12 columns must be in correct order, but additional (currently unused) columns are allowed
   includes code written by Gary Feng and Nigshan Zhang
}

