% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insertRow.R
\name{insertRow}
\alias{insertRow}
\title{insert a row at a specific position of a data frame}
\usage{
insertRow(existingDF, newrow, r)
}
\arguments{
\item{existingDF}{A data frame to add the row to}

\item{newrow}{The row of data to add}

\item{r}{The row number to insert the row to.}
}
\description{
insert a row at a specific position of a data frame
}

