% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_plot.R
\name{missing_states}
\alias{missing_states}
\title{Missing States}
\usage{
missing_states(allFips, stateZips)
}
\arguments{
\item{allFips}{A data frame containing county and
fip code data.}

\item{stateZips}{A data frame containing both US zip
and fip codes.}
}
\description{
A function that calculates the missing
 FIP codes and initializes them to 0 to cover every
 county in the state.
}
