% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg}
\alias{pkg}
\title{load a R package, download if necessary}
\usage{
pkg(x)
}
\arguments{
\item{x}{Name of the library to load, in string}
}
\description{
load a R package, download if necessary
}

