% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ppdvars.r
\docType{data}
\name{ppdvars}
\alias{ppdvars}
\title{Public Plans Database variables.}
\format{Data frame with 1 row per variable.}
\source{
Center for State and Local Government Excellence (SLGE),\cr
        National Association of State Retirement Administrators (NASRA), and\cr
        Center for Retirement Research at Boston College\cr\cr
        Link: http://publicplansdata.org/public-plans-database/download-full-data-set/\cr
}
\usage{
ppdvars
}
\description{
Metadata for the Public Plans Database.
}
\examples{
ppdvars
}
\keyword{datasets}

