% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_features.R
\name{ratio_features}
\alias{ratio_features}
\title{Ratio Features}
\usage{
ratio_features(map.data)
}
\arguments{
\item{map.data}{A data frame containing the full data
set, with a state column appended on.}
}
\description{
A function that takes a map.data data frame
 and creates ratios of typical features per state.
}
