% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renderHtmlWidgetList.R
\name{renderHtmlWidgetList}
\alias{renderHtmlWidgetList}
\title{Render a list of htmlWidgets using various tricks}
\usage{
renderHtmlWidgetList(widgetList, renderFunction)
}
\arguments{
\item{widgetList}{A list of htmlWidget objects to be rendered}

\item{renderFunction}{The function to render individual widgets. It can be either a name
of the rendering function, e.g., "render_graph" in DiagrammeR, or the actual function to
be passed to this call.}
}
\value{
The knitted string. This is to be included in the output by using `r renderHtmlWidgetList(...)`;
}
\description{
Render a list of htmlWidgets using various tricks
}
\details{
This is a collection of various tricks. See the URL citations in the code.
  Note that this code does alliterate global variables starting with "renderHtmlWidgetList_".
  You may want to delete them using rm(list = ls(pattern="renderHtmlWidgetList_*")).
}
\examples{
Inlcude the following in the Rmd directly
  `r require(DiagrammeR); renderHtmlWidgetList(grfDf$graph, render_graph)`
}

