% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_time.R
\name{reporting_time}
\alias{reporting_time}
\title{Reporting Time}
\usage{
reporting_time(map.data)
}
\arguments{
\item{map.data}{A data frame containing the full data
set, with a state column appended on.}
}
\description{
A function that takes a map.data data frame
 and creates boxplots of turnaround times for each particular
 US state.
}
