% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scanpathPlot.R
\name{scanpathPlot}
\alias{scanpathPlot}
\title{This function reads in a picture file specified by the user, then overlays it with a heat map computed according to user
  specifications.}
\usage{
scanpathPlot(dframe, pict, path = ".", outputtitle, dx = 0, dy = 60,
  alphaLevel = 0.4, arrows = T, pathAlpha = 0.8, pathSize = 0.7,
  sColor = "red", mColor = "green", eColor = "blue", x0 = -1, y0 = -1,
  x1 = -1, y1 = -1)
}
\arguments{
\item{dframe}{Data frame containing the data to be used, with at least $x, $y, and $dur}

\item{pict}{String giving the base name of the PNG file to be used, e.g., "background.png"}

\item{path}{String giving the path of the PNG file to be used, e.g., path=getwd()}

\item{outputtitle}{Optional argument, allowing the user to specify the output title}

\item{dx}{Optional argument, adjustment to x dimension, default to 0px}

\item{dy}{Optional argument, adjustment to y dimension, default to 60px}

\item{alphaLevel}{Optional argument, transparency level, default to 0.4}

\item{arrows}{Optional argument, whether to draw arrows or just fixations; default to T}

\item{pathAlpha}{Optional argument, transparency level for paths, default to 0.8}

\item{pathSize}{Optional argument, path size default to 0.7}

\item{sColor}{Optional argument, starting color, default to "red"}

\item{mColor}{Optional argument, median color, default to "green"}

\item{eColor}{Optional argument, end color, default to "blue"}
}
\value{
ggplot object: heat map or fixation map of gaze time overlayed onto a picture read from file
}
\description{
This function reads in a picture file specified by the user, then overlays it with a heat map computed according to user
  specifications.
}
\examples{
scanpathPlot(tmp, paste(storyName, ".png", sep=""), path=getwd(),
    alphaLevel = alphaLevel)
}

