% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepen.frb.r
\docType{data}
\name{statepen.frb}
\alias{statepen.frb}
\title{Federal Reserve Board state summaries.}
\format{Data frame with 1 row per state per variable per year.
\describe{
\item{stabbr}{State postal abbreviation, character}
\item{vname}{Variable name}
\item{year}{Calendar year, integer}
\item{value}{Value; dollar amounts are in $ millions, numeric}
}}
\source{
Federal Reserve Board https://www.federalreserve.gov/apps/fof/efa/efa-project-state-local-government-defined-benefit-pension-plans.htm
}
\usage{
statepen.frb
}
\description{
State and Local Government Defined Benefit Pension Plans: State-level Detail, annual calendar years 2000+
Liabilities and discounted at the following rates:
  2000-2003:  6.0%
  2004-2009:  5.5%
  2010-2014:  5.0%.
}
\examples{
library(dplyr)
glimpse(statepen.frb)
}
\keyword{datasets}

