% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trackStates.R
\name{trackStates}
\alias{trackStates}
\title{compute states from events}
\usage{
trackStates(df, stateVarName = "states")
}
\arguments{
\item{df}{Data frame from which to compute the state variables}

\item{stateVarName}{The name of the state variable to be exported. Not used in the current version
as we return just the states as a list of lists}
}
\value{
A list of lists that represent the state variables.
}
\description{
compute states from events
}
\details{
The current code is specifically for the Mango reading SBT task. We will find ways to
   abstract this and make it more general.
}

