% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenretss.r
\docType{data}
\name{cenretss}
\alias{cenretss}
\title{Census retirement state summaries.}
\format{
Data frame with 1 row per state per administration level per variable per year.
\describe{
\item{year}{Pension systems' fiscal year, integer}
\item{stabbr}{State postal abbreviation, character}
\item{admin}{Administration level: 1 (State-local), 2 (State), or 3 (Local), integer}
\item{adminf}{Administration level - factor: State-local, State, or Local, character}
\item{variable}{Variable name}
\item{value}{State-level summary of value; dollar amounts are in $ thousands, numeric}
}
}
\source{
U.S. Bureau of the Census
}
\usage{
cenretss
}
\description{
State-level summaries of Census Bureau Retirement Survey, selected years 1957-1977, annual 1982-2016.
}
\examples{
library(dplyr)
glimpse(cenretss)
}
\keyword{datasets}
