% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{map_theme}
\alias{map_theme}
\title{Map Theme}
\usage{
map_theme(fontFamily = "Gotham", titleFont = 26)
}
\arguments{
\item{fontFamily}{A character string that defines the font
family to be used for the axis / title / legend etc.
Default to _Gotham_.}

\item{titleFont}{An integer value that defines the title
font size.}
}
\description{
Holds the theme for the plots using
 ggplot choroplethr images.
}
