% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdata.r
\docType{package}
\name{pdata}
\alias{pdata}
\alias{pdata-package}
\title{Data for analysis of public pension plans.}
\description{
Currently includes:\cr
\code{\link{ppd}} Public Plans Database\cr
\code{\link{ppdvars}} Metadata for the Public Plans Database\cr
}

