% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readXML.R
\name{readXML}
\alias{readXML}
\title{Read a list of XML files and return a data frame containing all non eNAEP API events}
\usage{
readXML(xmlFiles, subjIdVar = "bookletId")
}
\arguments{
\item{xmlFiles}{A vector of eNAEP XML file names with full path;
typically from \code{Sys.glob("path/*")}}

\item{subjIdVar}{String of the name of the variable in the returned data frame for the
subjectId, which is the filename of the XML file; default to rewriting "bookletId".
If that's not appropriate, use another meaningful variable name.}
}
\description{
Read a list of XML files and return a data frame containing all non eNAEP API events
}
\examples{
test <- readXML(Sys.glob("data/*"), subjIdVar="subj")
}

